pam_setquota: fix more harmless compilation warnings
On ppc64le the compiler complains with the following diagnostics:

pam_setquota.c: In function 'debug':
pam_setquota.c:48:59: warning: format '%llu' expects argument of type 'long long unsigned int', but argument 6 has type '__u64' {aka 'const long unsigned int'} [-Wformat=]
   48 |   pam_syslog(pamh, LOG_DEBUG, "%s device=%s bsoftlimit=%llu bhardlimit=%llu "
      |                                                        ~~~^
      |                                                           |
      |                                                           long long unsigned int
      |                                                        %lu
......
   51 |              p->dqb_bsoftlimit, p->dqb_bhardlimit,
      |              ~~~~~~~~~~~~~~~~~
      |               |
      |               __u64 {aka const long unsigned int}
pam_setquota.c:48:75: warning: format '%llu' expects argument of type 'long long unsigned int', but argument 7 has type '__u64' {aka 'const long unsigned int'} [-Wformat=]
   48 |   pam_syslog(pamh, LOG_DEBUG, "%s device=%s bsoftlimit=%llu bhardlimit=%llu "
      |                                                                        ~~~^
      |                                                                           |
      |                                                                           long long unsigned int
      |                                                                        %lu
......
   51 |              p->dqb_bsoftlimit, p->dqb_bhardlimit,
      |                                 ~~~~~~~~~~~~~~~~~
      |                                  |
      |                                  __u64 {aka const long unsigned int}
pam_setquota.c:48:31: warning: format '%llu' expects argument of type 'long long unsigned int', but argument 8 has type '__u64' {aka 'const long unsigned int'} [-Wformat=]
   48 |   pam_syslog(pamh, LOG_DEBUG, "%s device=%s bsoftlimit=%llu bhardlimit=%llu "
      |                               ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
......
   52 |              p->dqb_isoftlimit, p->dqb_ihardlimit,
      |              ~~~~~~~~~~~~~~~~~
      |               |
      |               __u64 {aka const long unsigned int}
pam_setquota.c:49:46: note: format string is defined here
   49 |                               "isoftlimit=%llu ihardlimit=%llu btime=%llu itime=%llu",
      |                                           ~~~^
      |                                              |
      |                                              long long unsigned int
      |                                           %lu
pam_setquota.c:48:31: warning: format '%llu' expects argument of type 'long long unsigned int', but argument 9 has type '__u64' {aka 'const long unsigned int'} [-Wformat=]
   48 |   pam_syslog(pamh, LOG_DEBUG, "%s device=%s bsoftlimit=%llu bhardlimit=%llu "
      |                               ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
......
   52 |              p->dqb_isoftlimit, p->dqb_ihardlimit,
      |                                 ~~~~~~~~~~~~~~~~~
      |                                  |
      |                                  __u64 {aka const long unsigned int}
pam_setquota.c:49:62: note: format string is defined here
   49 |                               "isoftlimit=%llu ihardlimit=%llu btime=%llu itime=%llu",
      |                                                           ~~~^
      |                                                              |
      |                                                              long long unsigned int
      |                                                           %lu
pam_setquota.c:48:31: warning: format '%llu' expects argument of type 'long long unsigned int', but argument 10 has type '__u64' {aka 'const long unsigned int'} [-Wformat=]
   48 |   pam_syslog(pamh, LOG_DEBUG, "%s device=%s bsoftlimit=%llu bhardlimit=%llu "
      |                               ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
......
   53 |              p->dqb_btime, p->dqb_itime);
      |              ~~~~~~~~~~~~
      |               |
      |               __u64 {aka const long unsigned int}
pam_setquota.c:49:73: note: format string is defined here
   49 |                               "isoftlimit=%llu ihardlimit=%llu btime=%llu itime=%llu",
      |                                                                      ~~~^
      |                                                                         |
      |                                                                         long long unsigned int
      |                                                                      %lu
pam_setquota.c:48:31: warning: format '%llu' expects argument of type 'long long unsigned int', but argument 11 has type '__u64' {aka 'const long unsigned int'} [-Wformat=]
   48 |   pam_syslog(pamh, LOG_DEBUG, "%s device=%s bsoftlimit=%llu bhardlimit=%llu "
      |                               ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
......
   53 |              p->dqb_btime, p->dqb_itime);
      |                            ~~~~~~~~~~~~
      |                             |
      |                             __u64 {aka const long unsigned int}
pam_setquota.c:49:84: note: format string is defined here
   49 |                               "isoftlimit=%llu ihardlimit=%llu btime=%llu itime=%llu",
      |                                                                                 ~~~^
      |                                                                                    |
      |                                                                                    long long unsigned int
      |                                                                                 %lu

* modules/pam_setquota/pam_setquota.c (debug): Cast fields of type __u64
to unsigned long long.

+++ modules/pam_setquota/pam_setquota.c
@@ -47,10 +47,13 @@ debug(pam_handle_t *pamh, const struct if_dqblk *p,
       const char *device, const char *dbgprefix) {
   pam_syslog(pamh, LOG_DEBUG, "%s device=%s bsoftlimit=%llu bhardlimit=%llu "
                               "isoftlimit=%llu ihardlimit=%llu btime=%llu itime=%llu",
-             dbgprefix, device,
-             p->dqb_bsoftlimit, p->dqb_bhardlimit,
-             p->dqb_isoftlimit, p->dqb_ihardlimit,
-             p->dqb_btime, p->dqb_itime);
+	     dbgprefix, device,
+	     (unsigned long long) p->dqb_bsoftlimit,
+	     (unsigned long long) p->dqb_bhardlimit,
+	     (unsigned long long) p->dqb_isoftlimit,
+	     (unsigned long long) p->dqb_ihardlimit,
+	     (unsigned long long) p->dqb_btime,
+	     (unsigned long long) p->dqb_itime);
 }
 
 static unsigned long long
